requirements.txt
asgiref==3.3.1
certifi==2019.9.11
Django==3.1.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
pipenv==2018.11.26
pytz==2020.4
sqlparse==0.4.1
virtualenv==16.7.7
virtualenv-clone==0.5.3
